Split narrowing into its own file
