+++ src/bdiff_defs.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define buf_size 8192
+#define min_chunk_size 10
+#define max_chunk_size 10000
