+++ tests/unittest_narrowing.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void add_narrowing_test_funcs();
